fix a bunch of warnings from codebeat
