+++ gedbrowser-writer/src/main/java/org/schoellerfamily/gedbrowser/writer/GedWriter.java
@@ -8,6 +8,7 @@ import java.io.IOException;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.schoellerfamily.gedbrowser.datamodel.Root;
+import org.schoellerfamily.gedbrowser.writer.creator.GedWriterLineCreator;
 
 /**
  * @author Richard Schoeller
@@ -39,10 +40,7 @@ public class GedWriter {
         try (FileOutputStream fstream = new FileOutputStream(filename);
                 BufferedOutputStream bstream = new BufferedOutputStream(
                         fstream)) {
-            for (final GedWriterLine line : creator.getLines()) {
-                final String string = line.getLine();
-                bstream.write(string.getBytes());
-            }
+            writeTheLines(bstream);
         } catch (IOException e) {
             logger.error("Problem writing GEDCOM file", e);
         }
@@ -70,6 +68,20 @@ public class GedWriter {
                 return backupFile;
             }
             i++;
+        }
+    }
+
+    /**
+     * Loop through the lines from the line creator and write them to the stream.
+     *
+     * @param stream the stream to write to
+     * @throws IOException if there is a problem writing to the stream
+     */
+    private void writeTheLines(final BufferedOutputStream stream)
+            throws IOException {
+        for (final GedWriterLine line : creator.getLines()) {
+            final String string = line.getLine();
+            stream.write(string.getBytes());
         }
     }
