+++ gedbrowser-writer/src/main/java/org/schoellerfamily/gedbrowser/writer/creator/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Copyright 2017 Richard Schoeller
+ * Process the data model, producing line definitions for a GEDCOM file.
+ */
+package org.schoellerfamily.gedbrowser.writer.creator;
