+++ gedbrowser-writer/src/test/java/org/schoellerfamily/gedbrowser/writer/test/GedLineCreatorTest.java
@@ -25,7 +25,7 @@ import org.schoellerfamily.gedbrowser.datamodel.Trailer;
 import org.schoellerfamily.gedbrowser.datamodel.util.GedObjectBuilder;
 import org.schoellerfamily.gedbrowser.writer.GedWriterFile;
 import org.schoellerfamily.gedbrowser.writer.GedWriterLine;
-import org.schoellerfamily.gedbrowser.writer.GedWriterLineCreator;
+import org.schoellerfamily.gedbrowser.writer.creator.GedWriterLineCreator;
 
 /**
  * @author Dick Schoeller
